Force users to select a country
The HTML5 required attribute does not work on chosen selects. This is a
known bug since 2012: https://github.com/harvesthq/chosen/issues/515
The chosen project is dead, and all chosen selects will be replaced when
we upgrade bootstrap. For now, this fix works just as well in the self
registration form.