INVOICE: Yearly invoices job (#878)
* Remove unneeded rubocop:disable Metrics/MethodLength comment

* Abort if there is already a yearly invoices job running

* Implement active_members

* Use distinct for active members

Otherwise we'll get duplicated due to the left_outer_join.

* Only assert_no_other_job_running! on enqueue

To check it during perform we'd need to filter out the current job
running. Also it just protects from a manual perform call.

* Use 1.1. of the invoice year instead of the invoice date

* Actually call process_invoices in perform

* Working Spec

* Improve error handling

* Set ExternalInvoice.state to error when API response is bad

* Add progress reporting

* Add error handling

* Log job errors as HitobitoLogEntry

* Handle failure as well

* Cleanup CreateYearlyInvoicesJob

 * Use a helper for the reference date
 * Use enqueue hook instead of overriding enqueue!
 * Remove unneeded log messages
 * Use more generic rest client error
 * Extract clear_external_invoices into method.
 * Use class.name in other_job_running?

* Cleanup and fix progress calculation

Previously an additional query was done, which didn't count the members
that were actually processed, but just the erroneous ones.

* Use perform directly for good case job specs

Since we don't want to test the error handling of Delayed::Job in these,
we wan't rspec to be able to show the exceptions directly in case we
need to debug the spec.

* Test with multiple threads running

For this we can't match on the complete response anymore, since the
invoice_id's aren't deterministic anymore, due to race conditions in the
threads.

* Remove check for member.stammsektion_role

We filter the active_members so that they should always have a
stammsektion_role. If not it's an error and we should handle it.

* Extract submit_sales_orders from create_invoices

* Clear spurious invoices before performing job