Do not update primary group explicitely in membership mutations
fixes #1161

The primary group gets updated already in after_save/after_destroy
callbacks on Role, so we don't have to do it explicitely in the
membership mutations.

The previous implementation with `update!` naively assumed that the
updated people are valid, which is not necessarily the case so the user
got http 500 error responses.
For the membership mutation we should not care about person validity
unless it is a precondition for the mutation, then we should handle it
explicitely and should not try to update the primary group anyway.