Fix event participation table displays
Fixes #2089

Since #3250, inactive roles are taken into account for giving
SektionMemberAdmins access to the table display columns of people.

In #1868, this was extended to not crash for event participations.

Since hitobito/hitobito_sww#208, event_participations are polymorphic
and normal includes cannot be used for preloading the person anymore.

The quick fix is to abandon preloading the unscoped roles for
event_participations for now. It is not entirely clear to me whether
this past-role-based access was supposed to work for event
participations as well. If it was, this may need a different, more
extensive fix.