Create missing people managers on UndoTermination#save
`UndoTermination#save` should restore the family/household as it was
before terminating the membership. This change ensures that the people
managers get recreated, this was missing before.

Pragmatically we don't try to restore the people managers as they were
before the termiation, but we let `Household#create_missing_people_managers`
create new ones as valid for the current context/time.